[#52] Refactor: store.ts 분리 (#123)
