+++ fe/src/stores/NewIssuesSideStore.ts
@@ -1,17 +1,4 @@
-import { atom, selector } from 'recoil';
-
-// export const NewIssuesLabelIdAtom = atom<number[]>({
-//   key: 'NewIssuesLabelIdAtom',
-//   default: [],
-// });
-// export const NewIssuesAssigneeIdAtom = atom<number[]>({
-//   key: 'NewIssuesAssigneeIdAtom',
-//   default: [],
-// });
-// export const NewIssuesMilestoneIdAtom = atom<number[]>({
-//   key: 'NewIssuesMilestoneIdAtom',
-//   default: [],
-// });
+import { atom } from 'recoil';
 
 export type NewIssuesIdType = {
   labelList: number[];
